requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
pymodbus>=3.0.0
python-dateutil>=2.8.0
gevent>=24.2.1
gevent-websocket>=0.10.1
paho-mqtt>=1.6.0
# Phidgets support (install separately if needed)
Phidget22>=1.19.0
